LayoutTests/imported/w3c/web-platform-tests/html/semantics/the-button-element/command-and-commandfor/command-reflection.tentative-expected.txt



PASS invoker should reflect  properly
PASS invoker should reflect show-modal properly
PASS invoker should reflect toggle-popover properly
PASS invoker should reflect hide-popover properly
PASS invoker should reflect show-popover properly
PASS invoker should reflect close properly
PASS invoker should reflect --custom properly
FAIL invoker should reflect odd cased sHoW-MoDaL properly - as show-modal assert_equals: invoker should reflect odd cased sHoW-MoDaL properly - as show-modal expected "show-modal" but got "sHoW-MoDaL"
FAIL invoker should reflect odd cased tOgGlE-pOpOvEr properly - as toggle-popover assert_equals: invoker should reflect odd cased tOgGlE-pOpOvEr properly - as toggle-popover expected "toggle-popover" but got "tOgGlE-pOpOvEr"
FAIL invoker should reflect odd cased hIdE-pOpOvEr properly - as hide-popover assert_equals: invoker should reflect odd cased hIdE-pOpOvEr properly - as hide-popover expected "hide-popover" but got "hIdE-pOpOvEr"
FAIL invoker should reflect odd cased sHoW-pOpOvEr properly - as show-popover assert_equals: invoker should reflect odd cased sHoW-pOpOvEr properly - as show-popover expected "show-popover" but got "sHoW-pOpOvEr"
FAIL invoker should reflect odd cased ClOsE properly - as close assert_equals: invoker should reflect odd cased ClOsE properly - as close expected "close" but got "ClOsE"
PASS invoker should reflect odd cased --cUsToM properly - as --cUsToM
FAIL invoker should reflect the invalid value "invalid" as the empty string assert_equals: invoker should reflect the invalid value "invalid" as the empty string expected "" but got "invalid"
FAIL invoker should reflect the invalid value "show-invalid" as the empty string assert_equals: invoker should reflect the invalid value "show-invalid" as the empty string expected "" but got "show-invalid"
FAIL invoker should reflect the invalid value "foo-bar" as the empty string assert_equals: invoker should reflect the invalid value "foo-bar" as the empty string expected "" but got "foo-bar"
